refactor: split agent launch context setup
